Move RatesDocument into its own file
